requirements.txt
PyJWT>=0.1.6
Django>=1.3
djangorestframework>=2.3.10